requirements.txt
numpy
matplotlib
open3d
opencv-python
pillow
torch
torchvision
torchaudio
ultralytics
nuscenes-devkit
pandas
tqdm
pyquaternion
scipy
